flow all elements
