+++ src/entities/main-screen/bottom-row/bottom-row.module.scss
@@ -1,0 +1,1 @@
+@use "../../../styles/sass/variables.scss" as *;
