+++ src/entities/main-screen/screen.module.scss
@@ -4,4 +4,5 @@
   display: flex;
   flex-flow: column nowrap;
   gap: $size-normal;
+  padding: 0 $size-normal;
 }
