+++ src/entities/main-screen/top-row/top-row.module.scss
@@ -1,6 +1,9 @@
+@use "../../../styles/sass/variables.scss" as *;
+
 .block {
   display: flex;
-  flex-flow: row nowrap;
+  flex-flow: row wrap;
   justify-content: space-between;
+  gap: $size-normal;
   min-height: 25em;
 }
