+++ src/styles/original/style.scss
@@ -526,13 +526,6 @@ body {
   cursor: pointer;
 }
 
-#innerWrapper {
-  margin: 0;
-  padding: 0;
-  padding: 0.5%;
-  height: 100%;
-}
-
 .psText:hover,
 .bdHover:hover {
   text-decoration: underline;
